Added limited Database support
You can now insert data to to the database. The countdown feature is almost completely working (I forgot to add a small piece of code for it.) Updating alarm row data is now available through the settings popup window. No updating data for a watch just yet (not much to update right now in it's current form just yet) Next would be removing and loading data (changing between watches first.) Then we about where we were before and can add loading data on start of program soon after that possibly. Highlight color for watches on the watch menu on the left does not change as well. Finally database will need to know what watch is open at the current time which is not being updated.