Finished Design. Bugs (kinda) fixed
Design is finished with all windows completed designed. Changes might be made in future to the design if we want, but I think it looks good. I also put limits on what the user could enter for the total number of alarms and the name of watch (Too many characters in the watch name, a negative number for the total number of alarms, etc.)